GEMS-218: Add install dependency script and check build with ejdb2 (need cmake > cmake v3.24 )

+++ gemsfits-install-dependencies.sh
@@ -15,12 +15,14 @@ elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
 
 fi
 
-sudo rm -f /usr/local/lib/libnlopt.*
-sudo rm -f /usr/local/lib/libejdb.*
-sudo rm -f /usr/local/lib/libyaml-cpp.*
+#sudo rm -f /usr/local/lib/libnlopt.*
+#sudo rm -f /usr/local/lib/libejdb.*
+sudo rm -f /usr/local/lib/libejdb2.*
+#sudo rm -f /usr/local/lib/libyaml-cpp.*
 
 BUILD_TYPE=Release
 threads=3
+USING_EJDB=NO #$1 # OLD_EJDB
 
 
 # NLopt is a library for nonlinear local and global optimization
@@ -44,24 +46,45 @@ test -f /usr/local/lib/libnlopt.$EXTN || {
 }
 
 # EJDB2 is an embeddable JSON database engine.
-# git checkout v1.2.12
-test -f /usr/local/lib/libejdb.$EXTN || {
+if [ "$USING_EJDB" == "OLD_EJDB" ];
+  then
+  # git checkout v1.2.12
+  test -f /usr/local/lib/libejdb.$EXTN || {
 
+          # Building ejdb library
+          mkdir -p ~/code && \
+                  cd ~/code && \
+                  git clone https://github.com/Softmotions/ejdb.git -b  v1.2.12 && \
+                  cd ejdb && \
+                  mkdir -p build && \
+                  cd build && \
+                  cmake .. -DCMAKE_CXX_FLAGS=-fPIC -DCMAKE_BUILD_TYPE=Release && \
+                  make && \
+                  sudo make install
+
+          # Removing generated build files
+          cd ~ && \
+                   rm -rf ~/code
+  }
+  else
+  test -f /usr/local/lib/libejdb2.$EXTN || {
+
+        echo "Using ejdb2"
         # Building ejdb library
         mkdir -p ~/code && \
                 cd ~/code && \
-                git clone https://github.com/Softmotions/ejdb.git -b  v1.2.12 && \
+                git clone https://github.com/Softmotions/ejdb.git  && \
                 cd ejdb && \
                 mkdir -p build && \
                 cd build && \
-                cmake .. -DCMAKE_CXX_FLAGS=-fPIC -DCMAKE_BUILD_TYPE=Release && \
-                make && \
+                cmake ..  -DCMAKE_BUILD_TYPE=Release && \
                 sudo make install
 
         # Removing generated build files
         cd ~ && \
                  rm -rf ~/code
-}
+   }
+fi
 
 # YAMLCPP
 # https://github.com/jbeder/yaml-cpp/releases/tag/yaml-cpp-0.6.3
